pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["haspork <kondratev.pacha@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
Pillow = "^9.1.1"
django-environ = "^0.9.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.0"
django-grappelli = "^3.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
